+++ config/test.exs
@@ -1,3 +1,1 @@
 import Config
-
-config :bb_tui, command_timeout: 100